// backend/src/common/enums/urine-test.enums.ts
export enum CrystalType {
  URATOS_AMORFOS = 'URATOS AMORFOS',
  ACIDO_URICO = 'ACIDO URICO',
  OXALATOS_CALCIO_MONOHIDRATADO = 'OXALATOS DE CALCIO, MONOHIDRATADO',
  OXALATOS_CALCIO_DIHIDRATADO = 'OXALATOS DE CALCIO, DIHIDRATADO',
  CISTINA = 'CISTINA',
  COLESTEROL = 'COLESTEROL',
  TIROSINA = 'TIROSINA',
  ACIDO_HIPURICO = 'ACIDO HIPURICO',
  URATO_SODIO = 'URATO DE SODIO',
  LEUCINA = 'LEUCINA',
  SULFATO_CALCIO = 'SULFATO DE CALCIO',
  FOSFATO_TRIPLE = 'FOSFATO TRIPLE',
  FOSFATOS_AMORFOS = 'FOSFATOS AMORFOS',
  FOSFATO_CALCIO = 'FOSFATO DE CALCIO',
  CARBONATO_CALCIO = 'CARBONATO DE CALCIO',
  BIURATO_AMONIO = 'BIURATO DE AMONIO',
}

export enum CylinderType {
  HIALINOS = 'CILINDRO HIALINOS',
  ERITROCITARIO = 'CILINDRO ERITROCITARIO',
  LEUCOCITARIO = 'CILINDRO LEUCOCITARIO',
  GRANULOSO_FINO = 'CILINDRO GRANULOSO FINO',
  GRANULOSO_GRUESO = 'CILINDRO GRANULOSO GRUESO',
  CEREO = 'CILINDRO CEREO',
}

// Examen Físico
export enum UrineColor {
  AMARILLO = 'Amarillo',
  AMBAR = 'Ambar',
  CAFE = 'Café',
  ROJIZO = 'Rojizo',
  ROJO = 'Rojo',
  NARANJA = 'Naranja'
}

export enum UrineAspect {
  TRANSPARENTE = 'Transparente',
  LIGERAMENTE_TURBIO = 'Ligeramente Turbio',
  TURBIO = 'Turbio'
}

// Examen Químico
export enum UrineDensity {
  D1000 = '1.000',
  D1005 = '1.005',
  D1010 = '1.010',
  D1015 = '1.015',
  D1020 = '1.020',
  D1025 = '1.025',
  D1030 = '1.030'
}

export enum UrinePH {
  PH50 = '5.0',
  PH60 = '6.0',
  PH65 = '6.5',
  PH70 = '7.0',
  PH80 = '8.0',
  PH90 = '9.0'
}

export enum Urobilinogen {
  U01 = '0.1 mg/dl',
  U16 = '16 mg/dl',
  U33 = '33 mg/dl',
  U66 = '66 mg/dl',
  U131 = '131 mg/dl'
}

// Examen Microscópico
export enum NotObserved {
  NO_SE_OBSERVA = 'No se observa'
}